[KAN-92] refactor : 별점 계산 함수를 별도 파일로 분리
