+++ src/app/(main)/reviews/_components/ReviewScore/AverageRating.tsx
@@ -1,39 +1,10 @@
 import { IconHeart } from '@/public/icons';
 import { ReviewScoreType } from '@/types/data.type';
+import { calculateAverageRating } from './calculates';
 
 interface AverageRatingProps {
   ratingData: ReviewScoreType[];
 }
-const calculateAverageRating = (data: ReviewScoreType[]) => {
-  if (data.length <= 0) {
-    return '0';
-  }
-  let totalWeightedScore = 0; // 가중치가 적용된 평점 합계
-  let totalReviews = 0; // 총 평점 수
-
-  data.forEach((item) => {
-    const weightedScore =
-      item.oneStar * 1 +
-      item.twoStars * 2 +
-      item.threeStars * 3 +
-      item.fourStars * 4 +
-      item.fiveStars * 5;
-
-    const reviewCount =
-      item.oneStar +
-      item.twoStars +
-      item.threeStars +
-      item.fourStars +
-      item.fiveStars;
-
-    totalWeightedScore += weightedScore;
-    totalReviews += reviewCount;
-  });
-
-  // 평균을 계산하고 결과 반환
-  const average = totalReviews === 0 ? 0 : totalWeightedScore / totalReviews;
-  return average.toFixed(1);
-};
 
 const AverageRating = ({ ratingData }: AverageRatingProps) => {
   const rating = calculateAverageRating(ratingData); // 평균 점수
